+++ compositor_render/src/transformations.rs
@@ -1,7 +1,6 @@
 pub mod builtin;
 pub mod image_renderer;
 pub mod shader;
-mod shader_params;
 pub mod text_renderer;
 pub mod transition;
 pub mod web_renderer;
